corrected files for parsing
