+++ scripts/parsing/LeonBet_WC2022_short .py
@@ -10,7 +10,7 @@ logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s
 
 from fake_useragent import UserAgent
 
-import split_div_SportBet
+import split_div_LeonBet
 import save_to_csv
 
 user_agent = UserAgent(verify_ssl=False).chrome
@@ -29,26 +29,25 @@ chrome_options.add_argument("user-agent=" + user_agent)
 driver.implicitly_wait(10) # in seconds
 
 # ЧМ 2022
-driver.get("https://www.ligastavok.ru/bets/my-line/soccer/chm-2022-katar-id-1608/chm-2022-gruppovoi-etap-id-16538")
+driver.get("https://leon.ru/bets/soccer/world/1970324836989529-world-cup-2022")
 
 # wait until page will downloaded in browser
 time.sleep(15)
 
 try:    
     # page_source = driver.page_source
-    # print(f'page_source: {page_source}')
+    # print(page_source)
 
     # save screenshot of the page
-    driver.save_screenshot('LigaStavok_world_cup_2022.png')
+    driver.save_screenshot('LeonBet_world_cup_2022.png')
 
-    xp_block = '//*[@id="content"]/div/div/div[2]/div/section'
+    xp_block = '//*[@id="app"]/section/div/main/div/div/div[2]/div/div[3]/div[2]/div/div/div[2]'
     block_name = driver.find_elements(By.XPATH, xp_block)    
     block_name_list = [value.text for value in block_name]
-    print(f'\nblock_name_list: {block_name_list}\n')
+    # print(f'\nblock_name_list: {block_name_list}\n')
 
-    # splitted_list = split_div_LigaStavok.split_list(block_name_list) # split List
-    # save_to_csv.save_to_file(splitted_list, 'LigaStavok') # splitted List saving to csv-file
+    splitted_list = split_div_LeonBet.split_list(block_name_list) # split List
+    save_to_csv.save_to_file(splitted_list, 'LeonBet') # splitted List saving to csv-file
     
 except:
     print("some error happen !!")
-
