--- scripts/parsing/split_div_BaltBet copy.py
@@ -1,47 +1,0 @@
-import save_to_csv
-block_name_list = ['Катар — Эквадор\n20.11.22 19:00\n3.60\n3.25\n2.15\n1.70\n1.34\n1.29\n0\n2.52\n0\n1.54\n2,5\n2.31\n1.63\n609\nАнглия — Иран\n21.11.22 16:00\n1.31\n5.00\n11.50\n1.04\n1.17\n3.45\n-1,5\n2.01\n+1,5\n1.82\n2,5\n2.04\n1.80\n589\nСенегал — Нидерланды\n21.11.22 19:00\n6.00\n3.70\n1.62\n2.27\n1.27\n1.12\n+1\n1.77\n-1\n2.08\n2,5\n2.07\n1.78\n607\nСША — Уэльс\n21.11.22 22:00\n2.55\n3.05\n3.00\n1.38\n1.37\n1.51\n0\n1.77\n0\n2.08\n2,5\n2.47\n1.56\n604\nАргентина — Саудовская Аравия\n22.11.22 13:00\n1.16\n7.20\n19.00\n-\n1.09\n5.10\n-2\n1.89\n+2\n1.93\n2,5\n1.65\n2.27\n579\nДания — Тунис\n22.11.22 16:00\n1.45\n4.20\n7.90\n1.07\n1.22\n2.70\n-1\n1.77\n+1\n2.08\n2,5\n2.08\n1.77\n593\nМексика — Польша\n22.11.22 19:00\n2.75\n3.10\n2.75\n1.44\n1.36\n1.44\n0\n1.91\n0\n1.91\n2,5\n2.42\n1.58\n604\nФранция — Австралия\n22.11.22 22:00\n1.22\n6.20\n13.50\n1.02\n1.11\n4.20\n-1,5\n1.71\n+1,5\n2.17\n2,5\n1.65\n2.27\n593\nМарокко — Хорватия\n23.11.22 13:00\n4.60\n3.30\n1.87\n1.91\n1.32\n1.19\n+1\n1.47\n-1\n2.73\n2,5\n2.31\n1.63\n601\nГермания — Япония\n23.11.22 16:00\n1.40\n4.80\n7.70\n1.08\n1.18\n2.93\n-1,5\n2.17\n+1,5\n1.71\n2,5\n1.68\n2.22\n611\nИспания — Коста-Рика\n23.11.22 19:00\n1.20\n6.30\n17.00\n1.01\n1.12\n4.50\n-2\n2.15\n+2\n1.72\n2,5\n1.78\n2.07\n579\nБельгия — Канада\n23.11.22 22:00\n1.34\n5.10\n9.00\n1.06\n1.16\n3.20\n-1,5\n2.03\n+1,5\n1.81\n2,5\n1.71\n2.17\n606\nШвейцария — Камерун\n24.11.22 13:00\n1.77\n3.45\n5.10\n1.17\n1.31\n2.05\n-1\n2.50\n+1\n1.55\n2,5\n2.45\n1.57\n600\nУругвай — Южная Корея\n24.11.22 16:00\n1.77\n3.45\n5.00\n1.17\n1.30\n2.03\n-1\n2.50\n+1\n1.55\n2,5\n2.35\n1.61\n599\nПортугалия — Гана\n24.11.22 19:00\n1.45\n4.20\n8.30\n1.06\n1.22\n2.75\n-1\n1.75\n+1\n2.11\n2,5\n2.15\n1.72\n592\nБразилия — Сербия\n24.11.22 22:00\n1.46\n4.40\n7.20\n1.09\n1.21\n2.70\n-1\n1.77\n+1\n2.08\n2,5\n1.84\n1.99\n612\nУэльс — Иран\n25.11.22 13:00\n2.20\n3.15\n3.60\n1.28\n1.35\n1.65\n0\n1.54\n0\n2.52\n2,5\n2.45\n1.57\n598\nКатар — Сенегал\n25.11.22 16:00\n3.70\n3.25\n2.10\n1.72\n1.33\n1.27\n0\n2.66\n0\n1.49\n2,5\n2.29\n1.64\n612\nНидерланды — Эквадор\n25.11.22 19:00\n1.64\n3.90\n5.30\n1.15\n1.25\n2.23\n-1\n2.11\n+1\n1.75\n2,5\n1.82\n2.01\n623\nАнглия — США\n25.11.22 22:00\n1.65\n3.80\n5.40\n1.15\n1.26\n2.20\n-1\n2.20\n+1\n1.69\n2,5\n2.03\n1.81\n612\nТунис — Австралия\n26.11.22 13:00\n2.90\n3.00\n2.70\n1.45\n1.38\n1.40\n0\n2.00\n0\n1.83\n2,5\n2.55\n1.53\n602\nПольша — Саудовская Аравия\n26.11.22 16:00\n1.65\n3.80\n5.50\n1.15\n1.26\n2.23\n-1\n2.15\n+1\n1.72\n2,5\n2.04\n1.80\n604\nФранция — Дания\n26.11.22 19:00\n1.95\n3.35\n4.10\n1.23\n1.32\n1.83\n0\n1.43\n0\n2.90\n2,5\n2.27\n1.65\n601\nАргентина — Мексика\n26.11.22 22:00\n1.60\n3.80\n6.10\n1.12\n1.26\n2.33\n-1\n2.08\n+1\n1.77\n2,5\n2.15\n1.72\n601\nЯпония — Коста-Рика\n27.11.22 13:00\n1.90\n3.35\n4.40\n1.21\n1.32\n1.89\n-1\n2.85\n+1\n1.44\n2,5\n2.37\n1.60\n601\nБельгия — Марокко\n27.11.22 16:00\n1.52\n4.20\n6.30\n1.11\n1.22\n2.50\n-1\n1.90\n+1\n1.92\n2,5\n1.89\n1.93\n606\nХорватия — Канада\n27.11.22 19:00\n1.80\n3.60\n4.50\n1.20\n1.28\n1.99\n-1\n2.55\n+1\n1.53\n2,5\n2.10\n1.76\n609\nИспания — Германия\n27.11.22 22:00\n2.55\n3.35\n2.75\n1.44\n1.32\n1.50\n0\n1.84\n0\n1.99\n2,5\n1.93\n1.89\n622\nКамерун — Сербия\n28.11.22 13:00\n4.20\n3.30\n1.95\n1.84\n1.33\n1.22\n+1\n1.43\n-1\n2.90\n2,5\n2.45\n1.57\n600\nЮжная Корея — Гана\n28.11.22 16:00\n2.65\n2.95\n3.00\n1.38\n1.39\n1.47\n0\n1.79\n0\n2.05\n1,5\n1.52\n2.57\n601\nБразилия — Швейцария\n28.11.22 19:00\n1.50\n4.20\n6.70\n1.10\n1.22\n2.55\n-1\n1.85\n+1\n1.98\n2,5\n1.90\n1.92\n615\nПортугалия — Уругвай\n28.11.22 22:00\n2.10\n3.30\n3.70\n1.26\n1.32\n1.72\n0\n1.50\n0\n2.63\n2,5\n2.24\n1.67\n611\nЭквадор — Сенегал\n29.11.22 18:00\n2.90\n3.15\n2.55\n1.50\n1.35\n1.40\n0\n2.05\n0\n1.79\n2,5\n2.24\n1.67\n620\nНидерланды — Катар\n29.11.22 18:00\n1.33\n5.30\n9.00\n1.06\n1.15\n3.30\n-1,5\n2.00\n+1,5\n1.84\n2,5\n1.66\n2.25\n612\nУэльс — Англия\n29.11.22 22:00\n6.10\n3.90\n1.58\n2.35\n1.25\n1.12\n+1,5\n1.47\n-1,5\n2.75\n2,5\n1.97\n1.86\n617\nИран — США\n29.11.22 22:00\n3.45\n3.40\n2.15\n1.69\n1.30\n1.30\n0\n2.52\n0\n1.54\n2,5\n1.98\n1.85\n625\nАвстралия — Дания\n30.11.22 18:00\n6.40\n3.70\n1.60\n2.33\n1.28\n1.11\n+1,5\n1.41\n-1,5\n2.95\n2,5\n2.35\n1.61\n602\nТунис — Франция\n30.11.22 18:00\n14.00\n5.70\n1.24\n4.00\n1.14\n1.02\n+1,5\n2.05\n-1,5\n1.79\n2,5\n1.90\n1.92\n589\nСаудовская Аравия — Мексика\n30.11.22 22:00\n6.30\n4.10\n1.55\n2.45\n1.22\n1.11\n+1,5\n1.50\n-1,5\n2.60\n2,5\n1.95\n1.88\n617\nПольша — Аргентина\n30.11.22 22:00\n4.80\n4.00\n1.67\n2.17\n1.23\n1.17\n+1\n1.69\n-1\n2.20\n2,5\n1.93\n1.89\n623\nХорватия — Бельгия\n01.12.22 18:00\n3.90\n3.60\n1.93\n1.86\n1.29\n1.25\n0\n2.95\n0\n1.42\n2,5\n1.96\n1.87\n628\nКанада — Марокко\n01.12.22 18:00\n3.15\n3.35\n2.30\n1.62\n1.32\n1.36\n0\n2.31\n0\n1.63\n2,5\n2.22\n1.68\n620\nЯпония — Испания\n01.12.22 22:00\n9.50\n5.40\n1.31\n3.40\n1.15\n1.05\n+1,5\n1.86\n-1,5\n1.95\n2,5\n1.63\n2.31\n614\nКоста-Рика — Германия\n01.12.22 22:00\n11.00\n5.60\n1.28\n3.70\n1.14\n1.04\n+1,5\n1.98\n-1,5\n1.84\n2,5\n1.66\n2.25\n605\nГана — Уругвай\n02.12.22 18:00\n4.80\n3.60\n1.77\n2.05\n1.29\n1.18\n+1,5\n1.35\n-1,5\n3.30\n2,5\n2.08\n1.77\n618\nЮжная Корея — Португалия\n02.12.22 18:00\n9.50\n4.90\n1.35\n3.20\n1.18\n1.06\n+1,5\n1.71\n-1,5\n2.15\n2,5\n1.91\n1.91\n605\nСербия — Швейцария\n02.12.22 22:00\n2.75\n3.35\n2.55\n1.50\n1.32\n1.44\n0\n2.00\n0\n1.83\n2,5\n1.95\n1.88\n632\nКамерун — Бразилия\n02.12.22 22:00\n9.50\n5.30\n1.32\n3.37\n1.16\n1.05\n+1,5\n1.84\n-1,5\n1.98\n2,5\n1.68\n2.22\n614']
-
-def split_list(list):
-    result = []
-    for count, value in enumerate(list):
-        splitted = value.split('\\n')
-        splitted = value.split('\n')
-        # print(f'splitted:\n{splitted}')
-        separator = ':'        
-        how_many_games = [x for x in splitted if separator in x]
-        # print(f'\n\nhow_many_games: {how_many_games}')
-        # print(f'\n\nlen(how_many_games): {len(how_many_games)}')
-        start_search = 0 
-        # print(f'len(splitted): {len(splitted)}')
-        
-        for x in range(len(how_many_games)):
-            current_position = splitted.index(how_many_games[x], start_search)
-            # print(f'\nx: {x}')
-            # print(f'how_many_games[x]: {how_many_games[x]}')
-            # print(f'start_search: {start_search}')
-            # print(f'current_position: {current_position}')
-            index_of_date = splitted[current_position].index(':')  
-            date_of_game = splitted[current_position][:index_of_date-3]
-            time_of_game = splitted[current_position][index_of_date-2:] 
-            index_of_team = splitted[current_position - 1].index('—')
-            team_A = splitted[current_position - 1][:index_of_team-1]
-            team_B = splitted[current_position - 1][index_of_team+1:]       
-            result.append(['BaltBet',\
-                            'World cup 2022',\
-                            date_of_game,\
-                            time_of_game,\
-                            team_A,\
-                            team_B,\
-                            splitted[current_position + 1],\
-                            splitted[current_position + 2],\
-                            splitted[current_position + 3]
-                            ])
-            start_search = current_position + 1
-        
-    print(f'\nresult:\n {result}')
-    return result
-    
-# split_list(block_name_list)
-
-splitted_list = split_list(block_name_list) # split List
-save_to_csv.save_to_file(splitted_list, 'BaltBet') # splitted List saving to csv-file
