+++ scripts/parsing/split_div_BetCity.py
@@ -1,0 +1,52 @@
+# block_name_list = ['Футбол. Чемпионат мира 2022. Групповой этап. Катар.\n20 ноября\n19:00\nКатар\nЭквадор\nМатч ТВ\n3.80\n3.30\n2.18\n0\n2.65\n0\n1.52\n2.5\n1.60\n2.43\n+143\n21 ноября\n16:00\nАнглия\nИран\n1.32\n5.40\n12.0\n-1.5\n2.04\n+1.5\n1.83\n2.5\n1.81\n2.06\n+143\n19:00\nСенегал\nНидерланды\n6.60\n3.70\n1.66\n+1\n1.77\n-1\n2.12\n2.5\n1.76\n2.13\n+143\n22:00\nСША\nУэльс\n2.65\n3.20\n3.00\n0\n1.78\n0\n2.10\n2.5\n1.59\n2.45\n+144\n22 ноября\n13:00\nАргентина\nСаудовская Аравия\n1.16\n8.30\n23.0\n-2\n1.90\n+2\n1.96\n2.5\n2.36\n1.63\n+143\n16:00\nДания\nТунис\n1.42\n4.80\n8.80\n-1\n1.72\n+1\n2.20\n2.5\n1.85\n2.02\n+143\n19:00\nМексика\nПольша\n2.80\n3.20\n2.80\n0\n1.94\n0\n1.92\n2.5\n1.57\n2.50\n+143\n22:00\nФранция\nАвстралия\n1.23\n6.90\n15.0\n-1.5\n1.72\n+1.5\n2.20\n3\n1.80\n2.08\n+144\n23 ноября\n13:00\nМарокко\nХорватия\n4.80\n3.30\n1.93\n0\n3.40\n0\n1.35\n2.5\n1.63\n2.36\n+143\n16:00\nГермания\nЯпония\n1.44\n5.00\n7.70\n-1.5\n2.24\n+1.5\n1.70\n2.5\n2.26\n1.69\n+143\n19:00\nИспания\nКоста-Рика\n1.22\n6.80\n17.0\n-1.5\n1.69\n+1.5\n2.26\n2.5\n2.12\n1.77\n+143\n22:00\nБельгия\nКанада\n1.40\n5.10\n8.70\n-1\n1.59\n+1\n2.46\n2.5\n2.12\n1.77\n+143\n24 ноября\n13:00\nШвейцария\nКамерун\n1.78\n3.60\n5.40\n0\n1.29\n0\n3.90\n1.5\n2.75\n1.49\n+143\n16:00\nУругвай\nЮжная Корея\n1.77\n3.60\n5.40\n0\n1.28\n0\n3.90\n2.5\n1.63\n2.36\n+143\n19:00\nПортугалия\nГана\n1.46\n4.50\n8.40\n-1\n1.73\n+1\n2.18\n2.5\n1.74\n2.16\n+143\n22:00\nБразилия\nСербия\n1.44\n4.80\n8.00\n-1\n1.75\n+1\n2.16\n2.5\n2.10\n1.80\n+143\n25 ноября\n13:00\nУэльс\nИран\n2.23\n3.20\n3.60\n0\n1.56\n0\n2.55\n2.5\n1.60\n2.43\n+144\n16:00\nКатар\nСенегал\n3.60\n3.30\n2.20\n0\n2.55\n0\n1.55\n2.5\n1.63\n2.37\n+143\n19:00\nНидерланды\nЭквадор\n1.65\n4.10\n5.40\n-1\n2.14\n+1\n1.76\n2.5\n2.10\n1.79\n+143\n22:00\nАнглия\nСША\n1.65\n3.90\n5.80\n-1\n2.20\n+1\n1.72\n2.5\n1.83\n2.04\n+143\n26 ноября\n13:00\nТунис\nАвстралия\n3.00\n3.10\n2.65\n0\n2.03\n0\n1.84\n2.5\n1.54\n2.60\n+144\n16:00\nПольша\nСаудовская Аравия\n1.68\n3.90\n5.40\n-1\n2.30\n+1\n1.66\n2.5\n1.83\n2.04\n+143\n19:00\nФранция\nДания\n2.05\n3.40\n3.90\n0\n1.46\n0\n2.85\n2.5\n1.71\n2.22\n+143\n22:00\nАргентина\nМексика\n1.60\n4.00\n6.20\n-1\n2.13\n+1\n1.76\n2.5\n1.77\n2.12\n+143\n27 ноября\n13:00\nЯпония\nКоста-Рика\n1.90\n3.40\n4.70\n0\n1.34\n0\n3.40\n2.5\n1.65\n2.30\n+143\n16:00\nБельгия\nМарокко\n1.56\n4.30\n6.20\n-1\n2.00\n+1\n1.87\n2.5\n1.92\n1.94\n+143\n19:00\nХорватия\nКанада\n1.79\n3.90\n4.50\n-1\n2.60\n+1\n1.54\n2.5\n1.81\n2.07\n+143\n22:00\nИспания\nГермания\n2.60\n3.30\n2.85\n0\n1.84\n0\n2.02\n2.5\n1.94\n1.92\n+143\n28 ноября\n13:00\nКамерун\nСербия\n4.50\n3.30\n1.96\n0\n3.20\n0\n1.38\n2.5\n1.55\n2.55\n+142\n16:00\nЮжная Корея\nГана\n2.65\n3.10\n3.00\n0\n1.84\n0\n2.03\n2.5\n1.54\n2.60\n+143\n19:00\nБразилия\nШвейцария\n1.48\n4.40\n7.50\n-1\n1.83\n+1\n2.04\n2.5\n1.98\n1.88\n+143\n22:00\nПортугалия\nУругвай\n2.07\n3.40\n3.90\n0\n1.48\n0\n2.75\n2.5\n1.69\n2.24\n+143\n29 ноября\n18:00\nНидерланды\nКатар\n1.34\n5.00\n11.0\n-1.5\n1.90\n+1.5\n1.96\n2.5\n2.22\n1.71\n+144\n18:00\nЭквадор\nСенегал\n3.00\n3.20\n2.55\n0\n2.08\n0\n1.80\n2.5\n1.67\n2.28\n+144\n22:00\nУэльс\nАнглия\n6.10\n3.90\n1.62\n+1.5\n1.52\n-1.5\n2.65\n2.5\n1.83\n2.04\n+143\n22:00\nИран\nСША\n3.50\n3.40\n2.20\n0\n2.55\n0\n1.56\n2.5\n1.77\n2.12\n+144\n30 ноября\n18:00\nТунис\nФранция\n14.0\n5.80\n1.26\n+1.5\n2.08\n-1.5\n1.80\n2.5\n1.94\n1.92\n+144\n18:00\nАвстралия\nДания\n6.70\n3.70\n1.62\n+1.5\n1.47\n-1.5\n2.80\n2.5\n1.60\n2.43\n+143\n22:00\nПольша\nАргентина\n4.70\n4.30\n1.69\n0\n3.70\n0\n1.31\n2.5\n1.77\n2.12\n+144\n22:00\nСаудовская Аравия\nМексика\n6.50\n4.10\n1.56\n+1.5\n1.54\n-1.5\n2.55\n2.5\n1.82\n2.05\n+144\n1 декабря\n18:00\nХорватия\nБельгия\n4.00\n3.70\n1.94\n0\n3.00\n0\n1.43\n2.5\n1.83\n2.04\n+144\n18:00\nКанада\nМарокко\n3.30\n3.30\n2.32\n0\n2.33\n0\n1.65\n2.5\n1.70\n2.23\n+144\n22:00\nЯпония\nИспания\n12.5\n5.50\n1.29\n+1.5\n1.93\n-1.5\n1.93\n2.5\n2.32\n1.65\n+144\n22:00\nКоста-Рика\nГермания\n11.5\n5.10\n1.33\n+1.5\n1.98\n-1.5\n1.88\n2.5\n2.18\n1.73\n+143\n2 декабря\n18:00\nЮжная Корея\nПортугалия\n9.40\n4.80\n1.39\n+1\n2.32\n-1\n1.65\n2.5\n1.90\n1.96\n+144\n18:00\nГана\nУругвай\n4.90\n3.60\n1.80\n0\n3.60\n0\n1.32\n2.5\n1.73\n2.18\n+144\n22:00\nКамерун\nБразилия\n9.70\n5.40\n1.34\n+1.5\n1.86\n-1.5\n2.00\n2.5\n2.14\n1.76\n+144\n22:00\nСербия\nШвейцария\n2.85\n3.40\n2.55\n0\n2.06\n0\n1.82\n2.5\n1.83\n2.04\n+144']
+
+def split_list(list):
+    result = []
+    for count, value in enumerate(list):
+        splitted = value.split('\\n')
+        splitted = value.split('\n')
+        # print(f'splitted:\n{splitted}')        
+        separator = ':'
+        # how_many_games = splitted.count(separator)
+        how_many_games = [x for x in splitted if separator in x]
+        # print(f'\n\nhow_many_games: {how_many_games}')
+        # print(f'\n\nlen(how_many_games): {len(how_many_games)}')
+        start_search = 0 
+        # print(f'len(splitted): {len(splitted)}')        
+        for x in range(len(how_many_games)):
+            current_position = splitted.index(how_many_games[x], start_search)
+            # print(f'\nx: {x}')
+            # print(f'how_many_games[x]: {how_many_games[x]}')
+            # print(f'start_search: {start_search}')
+            # print(f'current_position: {current_position}')            
+            if start_search == 0:
+                result.append(['BetCity',\
+                                'World cup 2022',\
+                                splitted[current_position - 1],\
+                                splitted[current_position],\
+                                splitted[current_position + 1],\
+                                splitted[current_position + 2],\
+                                splitted[current_position + 4],\
+                                splitted[current_position + 5],\
+                                splitted[current_position + 6]
+                                ])
+            else:
+                result.append(['BetCity',\
+                                'World cup 2022',\
+                                '',\
+                                splitted[current_position],\
+                                splitted[current_position + 1],\
+                                splitted[current_position + 2],\
+                                splitted[current_position + 3],\
+                                splitted[current_position + 4],\
+                                splitted[current_position + 5]
+                                ])
+
+            start_search = current_position + 1
+        # print(f'\n\ncount:\n {count}\n value:\n{value}')
+        
+
+    # print(f'\nresult:\n {result}')
+    return result
+    
+# split_list(block_name_list)
